Change name, reorg, add factories table
